pytestDemo/requirements.txt
jsonpath>=0.82
pytest>=6.2.4
PyYAML>=5.4.1
requests>=2.25.1
jsonschema>=2.6.0
PyMySQL>=0.9.3
allure-pytest>=2.8.6
genson>=1.0.2

